Refactor _tests
