+++ src/tapis-app/Sidebar/Sidebar.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
-import renderComponent from './utils/testing';
+import renderComponent from 'utils/testing';
 import Sidebar from 'tapis-app/Sidebar';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
