+++ src/tapis-ui/_common/AppIcon/AppIcon.test.js
