+++ src/tapis-ui/_common/Icon/Icon.test.js
