--- src/tapis-ui/_tests/components/Login.test.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import configureStore from 'redux-mock-store';
-import '@testing-library/jest-dom/extend-expect';
-import renderComponent from '../utils/testing';
-import { Login } from 'tapis-ui/components';
-import tapisReduxStore from 'fixtures/tapis-redux.fixture';
-
-const mockStore = configureStore();
-
-describe('Login', () => {
-  it('renders Login component', () => {
-    const store = mockStore(tapisReduxStore);
-
-    const { getAllByText } = renderComponent(<Login />, store);
-    expect(getAllByText(/Username/).length).toEqual(1);
-  });
-});
