--- src/tapis-ui/_tests/utils/tapis.js
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-const login = (authParams) => {
-  return axios.request({
-    method: 'post',
-    url: '/oauth/app',
-    data: authParams,
-  });
-};
-
-export default login;
