+++ src/tapis-ui/components/jobs/JobLauncher/JobLauncher.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
-import renderComponent from '../../utils/testing';
+import renderComponent from 'utils/testing';
 import { JobLauncher } from 'tapis-ui/components/jobs';
 import { Jobs } from '@tapis/tapis-typescript';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
@@ -20,7 +20,7 @@ describe('JobLauncher', () => {
       execSystemId: 'tapisv3-exec'
     }
     const { getAllByTestId } = renderComponent(<JobLauncher initialValues={initialValues} />, store);
-    const input = getAllByTestId('appId');
+    const input: any = getAllByTestId('appId');
     expect(input[0].value).toEqual('SleepSeconds');
   });
 });
