+++ src/tapis-ui/components/jobs/JobsListing/JobsListing.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
-import renderComponent from '../../utils/testing';
+import renderComponent from 'utils/testing';
 import { JobsListing } from 'tapis-ui/components/jobs';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
