+++ src/tapis-ui/components/systems/SystemList/SystemList.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
-import renderComponent from '../../utils/testing';
+import renderComponent from 'utils/testing';
 import { SystemList } from 'tapis-ui/components/systems';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
 import { tapisSystem } from 'fixtures/systems.fixtures';
@@ -14,7 +14,7 @@ jest.mock('tapis-hooks/systems');
 describe('SystemList', () => {
   it('renders SystemList component', () => {
     const store = mockStore(tapisReduxStore);
-    useList.mockReturnValue({
+    (useList as jest.Mock).mockReturnValue({
       data: {
         result: [ tapisSystem ],
       },
